Rename proxy_protocol to header_proxy_protocol
